Split up the huge views file into modules in a package
