+++ candidates/tests/test_version_diffs.py
@@ -1,6 +1,6 @@
 from django.test import TestCase
 
-from ..diffs import get_version_diffs
+from candidates.views.diffs import get_version_diffs
 
 class TestVersionDiffs(TestCase):
