+++ candidates/views/__init__.py
@@ -1,0 +1,7 @@
+from .candidacies import *
+from .constituencies import *
+from .frontpage import *
+from .help import *
+from .parties import *
+from .people import *
+from .users import *
