+++ candidates/views/diffs.py
